% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseval-constructors.R
\name{add_imputation}
\alias{add_imputation}
\title{Add imputation model to a psdesign object}
\usage{
add_imputation(psdesign, imputation)
}
\arguments{
\item{psdesign}{A psdesign object}

\item{imputation}{An imputation object, such as \link{impute_parametric}}
}
\description{
Add imputation model to a psdesign object
}

