% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseval-constructors.R
\name{add_riskmodel}
\alias{add_riskmodel}
\title{Add imputation model to a psdesign object}
\usage{
add_riskmodel(psdesign, riskmodel)
}
\arguments{
\item{psdesign}{A psdesign object}

\item{risk}{A risk model object, such as \link{risk_logistic}}
}
\description{
Add imputation model to a psdesign object
}

