% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{empirical_TE}
\alias{empirical_TE}
\title{Compute the empirical Treatment Efficacy}
\usage{
empirical_TE(psdesign, t)
}
\arguments{
\item{psdesign}{An object of class \link{psdesign}}

\item{t}{Fixed time for time to event outcomes to compute TE. If missing, uses restricted mean survival.}
}
\description{
Compute the empirical Treatment Efficacy
}
