% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{expand_augdata}
\alias{expand_augdata}
\title{Expand augmented data using the integration function}
\usage{
expand_augdata(model, psdesign, D = 500)
}
\arguments{
\item{model}{Formula defining the risk model}

\item{psdesign}{An object of class \link{psdesign}, that contains at least 1 integration model}

\item{D}{The number of samples to take for the simulated annealing}
}
\description{
Expand augmented data using the integration function
}
\keyword{Internal}
