% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-test-data.R
\name{generate_example_data}
\alias{generate_example_data}
\title{Generate sample data used for testing}
\usage{
generate_example_data(n)
}
\arguments{
\item{n}{Integer, the sample size}
}
\description{
Generate sample data used for testing
}

