% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_bivnorm}
\alias{impute_bivnorm}
\title{Imputation models for the missing S(1)}
\usage{
impute_bivnorm(mu = c(0, 0), sd = c(1, 1), rho = 0.2, ...)
}
\arguments{
\item{mu,}{means of the pair of surrogates, missing one first}

\item{sd,}{standard deviations of the pair, missing one first}

\item{rho,}{the correlation between X1 and X2}

\item{...}{Arguments passed to \link{glm}}
}
\description{
Imputation models for the missing S(1)
}

