% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_nonparametric}
\alias{impute_nonparametric}
\title{Nonparametric imputation model for the missing S(1)}
\usage{
impute_nonparametric(formula, ...)
}
\arguments{
\item{formula}{Formula specifying the imputation model for the surrogate
under treatment. Generally the candidate surrogate will be on the left side
in the formula, and the BIP or BIPs will be on the right side. In this case the BIP and the S(1) must be categorical.}

\item{...}{Arguments passed to}
}
\description{
Nonparametric imputation model for the missing S(1)
}

