% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_parametric}
\alias{impute_parametric}
\title{Imputation models for the missing S(1)}
\usage{
impute_parametric(formula, distribution = gaussian, ...)
}
\arguments{
\item{formula}{Formula specifying the imputation model for the surrogate
under treatment}

\item{distribution}{Assumed distribution for the imputation model. Must be
compatible with the \code{family} argument of \link{glm}. Currenly only
Gaussian models are supported}

\item{...}{Arguments passed to \link{glm}}
}
\description{
Imputation models for the missing S(1)
}

