% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_semiparametric}
\alias{impute_semiparametric}
\title{Semiparametric imputation model using the location-scale model}
\usage{
impute_semiparametric(formula.location, formula.scale, ...)
}
\arguments{
\item{formula.location}{Formula specifying the imputation model for the location component of the surrogate
under treatment. Generally the candidate surrogate will be on the left side
in the formula, and the BIP or BIPs will be on the right side}

\item{formula.scale}{Formula specifying the imputation model for the scale component of the surrogate
under treatment. Generally the candidate surrogate will be on the left side
in the formula, and the BIP or BIPs will be on the right side}
}
\description{
Semiparametric imputation model using the location-scale model
}

