% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate_semiparametric}
\alias{integrate_semiparametric}
\title{Semiparametric integration model using the location-scale model}
\usage{
integrate_semiparametric(formula.location, formula.scale, ...)
}
\arguments{
\item{formula.location}{Formula specifying the integration model for the location component of the surrogate
under treatment. Generally the candidate surrogate will be on the left side
in the formula, and the BIP or BIPs will be on the right side}

\item{formula.scale}{Formula specifying the integration model for the scale component of the surrogate
under treatment. Generally the candidate surrogate will be on the left side
in the formula, and the BIP or BIPs will be on the right side}

\item{...}{Other parameters passed to \link{sp_locscale}}
}
\description{
Semiparametric integration model using the location-scale model
}

