% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseval-constructors.R
\name{+.ps}
\alias{+.ps}
\title{Modify a psdesign object by adding on new components.}
\usage{
\method{+}{ps}(p1, p2)
}
\description{
This operator allows you to add objects to a psdesign object, such as imputation models and risk models
}
\details{
If the first object is an object of class \code{psdesign}, you can add
the following types of objects, and it will return a modified psdesign
object.

\itemize{
  \item \code{imputation}: Add or replace imputation model
  \item \code{riskmodel}: Add or replace risk model
}
}

