% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.R
\name{print.psdesign}
\alias{print.psdesign}
\title{Concisely print information about a psdesign object}
\usage{
\method{print}{psdesign}(psdesign, digits = 3, sig.level = 0.05)
}
\arguments{
\item{psdesign}{A \link{psdesign} object}

\item{digits}{Number of significant digits to display}

\item{sig.level}{Significance level to use for computing bootstrapped confidence intervals}
}
\description{
Concisely print information about a psdesign object
}

