% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ps_bootstrap}
\alias{ps_bootstrap}
\title{Estimate parameters from a specified model using bootstrap resampling and estimated maximum likelihood}
\usage{
ps_bootstrap(n.boots = 200, progress.bar = TRUE, start = NULL,
  method = "BFGS", control = list(), ...)
}
\arguments{
\item{n.boots}{Number of bootstrap replicates}

\item{progress.bar}{Logical, if true will display a progress bar in the console}

\item{start}{Vector of starting values, if NULL, will come up with starting values}

\item{method}{Method to use for optimization, can be "pseudo-score" for
categorical S with nonparametric integration, or any of the methods
available in \link{optim}. Defaults to "BFGS"}

\item{control}{List of control parameters for passed to \link{optim}}

\item{...}{Arguments passed to \link{optim}}
}
\description{
Estimate parameters from a specified model using bootstrap resampling and estimated maximum likelihood
}

