% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk.logit}
\alias{risk.logit}
\title{Logit link function}
\usage{
risk.logit(x)
}
\arguments{
\item{x}{A vector of linear predictors}
}
\value{
A vector of probabilities
}
\description{
Logit link function
}
