% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk.probit}
\alias{risk.probit}
\title{Probit link function}
\usage{
risk.probit(x)
}
\arguments{
\item{x}{A vector of linear predictors}
}
\value{
A vector of probabilities
}
\description{
Probit link function
}
