% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_continuous}
\alias{risk_continuous}
\title{Risk model for continuous outcome}
\usage{
risk_continuous(model = Y ~ S.1 * Z, D = 5000)
}
\arguments{
\item{model}{Formula specifying the risk model for the mean}

\item{D}{number of samples for the simulated annealing integration}
}
\description{
This model assumes that the outcome Y is normally distributed conditional on
S.1 and Z, with mean determined by the model formula. It also assumes that
larger values of Y are more indicative of poor outcomes, e.g., blood
pressure.
}
