% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_exponential}
\alias{risk_exponential}
\title{Exponential risk model for time to event outcome}
\usage{
risk_exponential(model = Y ~ S.1 * Z, D = 5000)
}
\arguments{
\item{model}{Formula specifying the risk model. The outcome should be a \link{Surv} object specifying right censoring}

\item{D}{number of samples for simulated annealing}
}
\description{
Exponential risk model for time to event outcome
}

