% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_logistic}
\alias{risk_logistic}
\title{Logistic risk model for binary outcome}
\usage{
risk_logistic(model = Y ~ S.1 * Z, D = 5000, ...)
}
\arguments{
\item{model}{Formula specifying the risk model}

\item{D}{number of samples for the simulated annealing integration}
}
\description{
Logistic risk model for binary outcome
}

