% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_poisson}
\alias{risk_poisson}
\title{Poisson risk model for count outcomes}
\usage{
risk_poisson(model = Y ~ S.1 * Z, D = 5000)
}
\arguments{
\item{model}{Formula specifying the risk model. The outcome should be an integer of counts. This right side of the formula may contain an \link{offset} term.}

\item{D}{number of samples for simulated annealing}
}
\description{
Poisson risk model for count outcomes
}

