% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_weibull}
\alias{risk_weibull}
\title{Weibull risk model for time to event outcome}
\usage{
risk_weibull(model = Y ~ S.1 * Z, cens = delt, D = 5000, ...)
}
\arguments{
\item{model}{Formula specifying the risk model}

\item{cens}{Symbol identifying the censoring indicator}

\item{D}{number of samples for simulated annealing}

\item{...}{}
}
\description{
Weibull risk model for time to event outcome
}

