% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{riskcalc}
\alias{riskcalc}
\title{Calculate risks with handlers for survival data}
\usage{
riskcalc(risk.function, Y, par, t, dat0, dat1)
}
\arguments{
\item{risk.function}{Function taking three arguments, a data.frame, parameters, and time. It should return a vector the same number of rows as the data frame}

\item{Y}{The outcome variable}

\item{par}{the vector of parameter values}

\item{t}{Time for a survival outcome, may be missing}

\item{dat0}{Data frame containing S and Z = 1}

\item{dat1}{Data frame containing S and Z = 0}
}
\description{
Calculate risks with handlers for survival data
}
\keyword{Internal}

