% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-estimation.R
\name{stg}
\alias{stg}
\title{Compute the standardized total gain}
\usage{
stg(R1, R0, S)
}
\arguments{
\item{R1}{Risk in the treatment group}

\item{R0}{Risk in the control group}

\item{S}{Values of S at which risk is calculated}
}
\description{
Compute the standardized total gain
}

