% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_bs}
\alias{summarize_bs}
\title{Summarize bootstrap samples}
\usage{
summarize_bs(bootdf, sig.level = 0.05)
}
\arguments{
\item{bootdf}{Data frame containing bootstrapped estimates, with a column containing a convergence indicator}

\item{sig.level}{Significance level to use for confidence intervals}
}
\description{
Summarize bootstrap samples
}

