% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{wem_test}
\alias{wem_test}
\title{Test for wide effect modificiation}
\usage{
wem_test(x)
}
\arguments{
\item{x}{An object of class \link{psdesign} with bootstrap replicates}
}
\description{
This runs a multivariate Wald test on the interaction terms of the model, using the bootstrap covariance
}

