% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseval-constructors.R
\name{add_estimate}
\alias{add_estimate}
\title{Estimate parameters}
\usage{
add_estimate(psdesign, estimate)
}
\arguments{
\item{psdesign}{A psdesign object, it must have risk model and integration model components}

\item{estimate}{An estimate object created by \link{ps_estimate}}
}
\description{
Estimate parameters
}
\examples{
test <- psdesign(generate_example_data(n = 100), Z = Z, Y = Y.obs, S = S.obs, BIP = BIP)
test + integrate_parametric(S.1 ~ BIP) + risk_binary(D = 50) + ps_estimate(method = "BFGS")
}
