% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{empirical_VE}
\alias{empirical_VE}
\title{Compute the empirical Treatment Efficacy}
\usage{
empirical_VE(psdesign, t)
}
\arguments{
\item{psdesign}{An object of class \link{psdesign}}

\item{t}{Fixed time for time to event outcomes to compute TE. If missing, uses restricted mean survival.}
}
\description{
Included for backwards compatibility
}
