% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{plot.psdesign}
\alias{plot.psdesign}
\title{Plot summary statistics for a psdesign object}
\usage{
\method{plot}{psdesign}(psdesign, t, summary = "VE", sig.level = 0.05, ...)
}
\arguments{
\item{psdesign}{A psdesign object that contains a risk model, imputation
model, and valid estimates}

\item{t}{For time to event outcomes, a fixed time \code{t} may be provided to
compute the cumulative distribution function. If not, the restricted mean
survival time is used. Omit for binary outcomes.}

\item{summary}{Summary statistic to plot. Currently only \link{VE} is
supported.}

\item{sig.level}{Significance level used for confidence bands on the VE
curve. This is only used if bootstrapped estimates are available.}

\item{...}{Other arguments passes to \link{plot}}
}
\description{
Plot the vaccine efficacy versus S.1 for an estimated psdesign object
}

