% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseval-constructors.R
\name{+.ps}
\alias{+.ps}
\title{Modify a psdesign object by adding on new components.}
\usage{
\method{+}{ps}(p1, p2)
}
\description{
This operator allows you to add objects to a psdesign object, such as integration models and risk models
}
\details{
If the first object is an object of class \code{psdesign}, you can add
the following types of objects, and it will return a modified psdesign
object. Users will generally add them in the order that they appear.

\itemize{
  \item \code{integration}: Add or replace integration model
  \item \code{riskmodel}: Add or replace risk model
  \item \code{estimate}: Estimate parameters
  \item \code{bootstrap}: Bootstrap estimates
}
}

