% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{ps_estimate}
\alias{ps_estimate}
\title{Estimate parameters from a specified model using estimated maximum likelihood}
\usage{
ps_estimate(psdesign, start = NULL, control = list(), ...)
}
\arguments{
\item{psdesign}{A psdesign object with a risk model and integration model specified}

\item{start}{Vector of starting values, if NULL, will come up with starting values}

\item{control}{List of control parameters for passed to \link{optim}}

\item{...}{Arguments passed to \link{optim}}
}
\description{
Estimate parameters from a specified model using estimated maximum likelihood
}

