% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/design.R
\name{psdesign}
\alias{psdesign}
\title{Specify a design for a principal surrogate evaluation}
\usage{
psdesign(data = NULL, treatment, outcome, surrogate, BIP = NULL,
  CPV = NULL, BSM = NULL, weights = NULL, ...)
}
\arguments{
\item{data}{Data frame containing data to be analyzed}

\item{treatment}{Expression defining the treatment variable}

\item{outcome}{Expression defining the outcome variable}

\item{surrogate}{Expression defining the candidate surrogate}

\item{BIP}{Optional expression defining the baseline irrelevant predictor}

\item{CPV}{Optional expression defining the closeout placebo vaccination measurement}

\item{BSM}{Optional expression defining the baseline surrogate measurement}

\item{weights}{optional expression defining weights to accomodate nonrandom subsampling, such as case control or two phase}

\item{...}{Not currently used}
}
\description{
Specify a design for a principal surrogate evaluation
}

