% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-estimation.R
\name{stg}
\alias{stg}
\title{Compute the standardized total gain}
\usage{
stg(R1, R0, stand = TRUE)
}
\arguments{
\item{R1}{Risk in the treatment group}

\item{R0}{Risk in the control group}

\item{stand}{Standardize?}
}
\description{
Compute the standardized total gain
}
