% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{summary.psdesign}
\alias{summary.psdesign}
\title{Summary method for psdesign objects}
\usage{
\method{summary}{psdesign}(psdesign, digits = 3, sig.level = 0.05)
}
\arguments{
\item{psdesign}{A \link{psdesign} object}

\item{digits}{Number of significant digits to display}

\item{sig.level}{Significance level to use for computing bootstrapped confidence intervals}
}
\description{
Summary method for psdesign objects
}

