% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.R
\name{VE}
\alias{VE}
\title{Calculate the vaccine efficacy}
\usage{
VE(psdesign, t)
}
\arguments{
\item{psdesign}{A psdesign object. It must contain a risk model, an
imputation model, and estimated parameters. Bootstrapped parameters are
optional}

\item{t}{For time to event outcomes, a fixed time \code{t} may be provided to
  compute the cumulative distribution function. If not, the restricted mean
  survival time is used. Omit for binary outcomes.}
}
\value{
A data frame containing columns for the S values, the VE at those S
  values, and optionally standard errors computed using bootstrapped
  estimates.
}
\description{
Computes the vaccince efficacy (VE) over the range of surrogate values
observed in the data. VE(s) is defined as 1 - risk(s, z = 1)/risk(s, z = 0),
where z is the treatment indicator. If any other variables are present in the
risk model, then the risk is computed at their median value.
}

